Move syncing test-openshift.yaml as soon as the file is in dist-git
What is the reason for moving updating test-openshift.yaml in to the dist-git part?
Before this fix, test-openshift.yaml was updated in upstream repository.
But in case of more versions, first update replace
variables like VERSION_NUMBER, OS_NUMBER, CONTAINER_NAME
with the first version.
For second version the test-openshift.yaml file is already
changed to the first version and template names like
VERSION_NUMBER, OS_NUMBER, CONTAINER_NAME do no exist
and instead of them the variables replaced in first version are used.
In this case the second version do not use proper variables, but
variables from first version.
This fix update variables in dist-git and upstream file is not touched
and contain template names for all versions.

Signed-off-by: Petr "Stone" Hracek <phracek@redhat.com>